requirements.txt
Flask>=2.1.2
pandas>=1.4.3
Flask_SQLAlchemy>=3.0.2